Fix more tests for CockroachDB (#4729)
