+++ test/integration2/schema/foreign-keys.spec.js
@@ -14,9 +14,12 @@ describe('Schema', () => {
     getAllDbs().forEach((db) => {
       describe(db, () => {
         let knex;
-        before(() => {
+        before(async () => {
           sinon.stub(Math, 'random').returns(0.1);
           knex = getKnexForDb(db);
+          await knex.schema.dropTableIfExists('foreign_keys_table_one');
+          await knex.schema.dropTableIfExists('foreign_keys_table_two');
+          await knex.schema.dropTableIfExists('foreign_keys_table_three');
         });
 
         after(() => {
@@ -40,10 +43,9 @@ describe('Schema', () => {
         });
 
         afterEach(async () => {
-          await knex.schema
-            .dropTable('foreign_keys_table_one')
-            .dropTable('foreign_keys_table_two')
-            .dropTable('foreign_keys_table_three');
+          await knex.schema.dropTable('foreign_keys_table_one');
+          await knex.schema.dropTable('foreign_keys_table_two');
+          await knex.schema.dropTable('foreign_keys_table_three');
         });
 
         describe('createForeignKey', () => {
@@ -261,6 +263,107 @@ describe('Schema', () => {
               `PRAGMA foreign_key_list('foreign_keys_table_one');`
             );
             expect(fks.length).to.equal(1);
+          });
+        });
+
+        describe('Schema (Foreign keys)', () => {
+          beforeEach(async () => {
+            await knex.schema
+              .dropTableIfExists('foreign_keys_table_one')
+              .dropTableIfExists('foreign_keys_table_four');
+
+            await knex.schema
+              .createTable('foreign_keys_table_four', (table) => {
+                table.string('col1').notNull();
+                table.string('col2').notNull();
+                table.primary(['col1', 'col2']);
+              })
+              .createTable('foreign_keys_table_one', (table) => {
+                table.increments();
+                table.integer('fkey_two').unsigned().notNull();
+                table
+                  .foreign('fkey_two')
+                  .references('foreign_keys_table_two.id');
+                table.string('fkey_four_part1');
+                table.string('fkey_four_part2');
+                table
+                  .foreign(['fkey_four_part1', 'fkey_four_part2'])
+                  .references(['col1', 'col2'])
+                  .inTable('foreign_keys_table_four');
+                table.integer('fkey_three').unsigned().notNull();
+                table
+                  .foreign('fkey_three')
+                  .references('foreign_keys_table_three.id')
+                  .withKeyName('fk_fkey_threeee');
+              });
+          });
+          afterEach(async () => {
+            await knex.schema.dropTableIfExists('foreign_keys_table_four');
+          });
+
+          describe('drop foreign key', () => {
+            it('correctly drops foreign key', async () => {
+              await knex('foreign_keys_table_two').insert({});
+              await knex('foreign_keys_table_three').insert({});
+              await knex('foreign_keys_table_four').insert({
+                col1: 'a',
+                col2: 'b',
+              });
+              const tableTwoEntry = await knex(
+                'foreign_keys_table_two'
+              ).select();
+              const tableThreeEntry = await knex(
+                'foreign_keys_table_three'
+              ).select();
+              await knex('foreign_keys_table_one').insert({
+                fkey_two: tableTwoEntry[0].id,
+                fkey_three: tableThreeEntry[0].id,
+                fkey_four_part1: 'a',
+                fkey_four_part2: 'b',
+              });
+              try {
+                await knex('foreign_keys_table_one').insert({
+                  fkey_two: 9999,
+                  fkey_three: 99,
+                  fkey_four_part1: 'a',
+                  fkey_four_part2: 'b',
+                });
+                throw new Error("Shouldn't reach this");
+              } catch (err) {
+                if (isSQLite(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into \`foreign_keys_table_one\` (\`fkey_four_part1\`, \`fkey_four_part2\`, \`fkey_three\`, \`fkey_two\`) values ('a', 'b', 99, 9999) - SQLITE_CONSTRAINT: FOREIGN KEY constraint failed`
+                  );
+                }
+                if (isPostgreSQL(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into "foreign_keys_table_one" ("fkey_four_part1", "fkey_four_part2", "fkey_three", "fkey_two") values ($1, $2, $3, $4) - insert or update on table "foreign_keys_table_one" violates foreign key constraint "foreign_keys_table_one_fkey_two_foreign"`
+                  );
+                }
+                if (isCockroachDB(knex)) {
+                  expect(err.message).to.equal(
+                    `insert into "foreign_keys_table_one" ("fkey_four_part1", "fkey_four_part2", "fkey_three", "fkey_two") values ($1, $2, $3, $4) - insert on table "foreign_keys_table_one" violates foreign key constraint "foreign_keys_table_one_fkey_two_foreign"`
+                  );
+                }
+                expect(err.message).to.include('constraint');
+              }
+
+              await knex.schema.alterTable(
+                'foreign_keys_table_one',
+                (table) => {
+                  table.dropForeign(['fkey_two']);
+                  table.dropForeign([], 'fk_fkey_threeee');
+                  table.dropForeign(['fkey_four_part1', 'fkey_four_part2']);
+                }
+              );
+
+              await knex('foreign_keys_table_one').insert({
+                fkey_two: 999,
+                fkey_three: 999,
+                fkey_four_part1: 'e',
+                fkey_four_part2: 'f',
+              });
+            });
           });
         });
       });
